% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_function.R
\name{CorrMatrix}
\alias{CorrMatrix}
\title{CorrMatrix Function}
\usage{
CorrMatrix(year)
}
\arguments{
\item{year}{select a given year.}
}
\description{
This function allows you create a correlation matrix for all numeric variables in the data for that year.
}
\examples{
CorrMatrix()
}
\keyword{corr}
