% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_function.R
\name{OldestPlayer}
\alias{OldestPlayer}
\title{OldestPlayer Function}
\usage{
OldestPlayer(year)
}
\arguments{
\item{year}{select a given year.}
}
\description{
This function allows you to find the oldest player in the NBA data for that year.
}
\examples{
OldestPlayer()
}
\keyword{oldest}
