% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_function.R
\name{SomePlayer}
\alias{SomePlayer}
\title{SomePlayer function}
\usage{
SomePlayer(year)
}
\arguments{
\item{year}{select a given year.}
}
\description{
This function allows you to generate a data point of interest in the dataset for some player
}
\examples{
SomePlayer()
}
\keyword{some}
